home3 done manybe

+++ src/main/java/com/test/effectivejava/pojo/Movie.java
@@ -13,5 +13,5 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class Movie {
     private String name;
-    private double rating;
+    private BigDecimal rating;
 }
